Update: Moved Footer, ViewMyWorkButton, WelcomScreen under component/HomePage directory
