+++ src/components/HomePage/WelcomeScreen/__tests__/WelcomeScreen.test.tsx
