+++ src/components/HomePage/WelcomeScreen/helpers.ts
