+++ src/components/HomePage/WelcomeScreen/index.ts
