+++ src/components/HomePage/WelcomeScreen/styles.css
